refactor: 상품 관리 컴포넌트 분리
