+++ src/origin/pages/CartPage.tsx
@@ -9,22 +9,20 @@ export const CartPage = () => {
     <div className="container mx-auto p-4">
       <h1 className="text-3xl font-bold mb-6">장바구니</h1>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <CartProvider>
-          <div>
-            <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
-            <div className="space-y-2">
-              <ProductList />
-            </div>
+        <div>
+          <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
+          <div className="space-y-2">
+            <ProductList />
           </div>
-          <div>
-            <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
-            <div className="space-y-2">
-              <CartItemList />
-              <CouponApplySection />
-              <CartSummary />
-            </div>
+        </div>
+        <div>
+          <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
+          <div className="space-y-2">
+            <CartItemList />
+            <CouponApplySection />
+            <CartSummary />
           </div>
-        </CartProvider>
+        </div>
       </div>
     </div>
   );
